% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexwidget.R
\name{flexwidget-shiny}
\alias{flexwidget-shiny}
\alias{flexwidgetOutput}
\alias{renderflexwidget}
\title{Shiny bindings for flexwidget}
\usage{
flexwidgetOutput(outputId, width = "100\%", height = "400px")

renderflexwidget(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a flexwidget}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using flexwidget within Shiny
applications and interactive Rmd documents.
}
\details{
If using CDN for dependencies, you can add `flexwidget_src` call
into the UI. If accepting input that affects the javascript being passed to the
widget, use caution as a user could inject something malicious. It is
probably best to only have the data input to the widget change from user
input.
}
