% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexwidget.R
\name{flexwidget}
\alias{flexwidget}
\title{Empty Widget for Custom Javascript}
\usage{
flexwidget(data = NULL, js_code = NULL, inner_html = NULL,
  resize = NULL, width = NULL, height = NULL, elementId = NULL,
  dependencies = NULL)
}
\arguments{
\item{data}{data to pass to the js_code}

\item{js_code}{javascript code, should be a function that takes in single
argument (the data)}

\item{inner_html}{html to insert into the widget div}

\item{resize}{javascript code for resizing, should be a function that takes
in width and height argument.}

\item{width}{width of widget, optional}

\item{height}{height of widget, optional}

\item{elementId}{id for widget, optional}

\item{dependencies}{javascript dependencies, should be list of html_dependency
objects created from \code{\link[htmltools]{htmlDependency}}}
}
\value{
an htmlwidget
}
\description{
The htmlwidgets package allows you to create R interfaces to javascript
visualization code. The flexwidget package is a very basic widget that on
its own does very little. Rather it sets up the widget so  you can pass in
custom javascript and html. See the vignette and examples for an illustration.
}
\examples{
# Trivial example: creates an h1 div. passes the data multiplied by 2 to that div.
flexwidget(data = 2,
  js_code = "function(data){document.getElementById('mydiv').textContent = data * 2}",
  inner_html = tags$h1(id = "mydiv"))

# Resize example
flexwidget(
  resize = "function(width, height){document.getElementById('mydiv').textContent = width}",
  inner_html = tags$h1("Resize this window to get width", id = "mydiv"))

# For more exciting examples see vignette.

}
