% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdn.R
\name{flexwidget_dep}
\alias{flexwidget_dep}
\alias{flexwidget_src}
\alias{flexwidget_dep}
\alias{flexwidget_css}
\title{Make tag list for javascript or css dependencies}
\usage{
flexwidget_src(...)

flexwidget_css(...)
}
\arguments{
\item{...}{urls for javascript or css CDNs}
}
\value{
a tagList with script or style tags for each input URL

\code{\link[htmltools]{tagList}}
}
\description{
Pass CDN urls to knitr document or shiny
}
\examples{

flexwidget_src("https://d3js.org/d3.v4.min.js",
               "http://dimplejs.org/dist/dimple.v2.3.0.min.js")

}
